Handle an MQTT error code that is not an issue.
When a network switch happens, like when you toggle flight mode on and off, the mqtt library can do a reconnect while while we are also triggering a reconnect. An error is returned stating a connection is in progress. We can safely ignore this error and not fail and stop the logging.